Added missing typecast.

+++ coders/heic.c
@@ -159,7 +159,7 @@ static inline void HEICSecurityLimits(const ImageInfo *image_info,
   height_limit=(int) MagickMin(GetMagickResourceLimit(WidthResource),INT_MAX);
   if (width_limit != INT_MAX || height_limit != INT_MAX)
     security_limits->max_image_size_pixels=(uint64_t) width_limit*height_limit;
-  max_profile_size=MagickMin(GetMaxProfileSize(),INT_MAX);
+  max_profile_size=(int) MagickMin(GetMaxProfileSize(),INT_MAX);
   if (max_profile_size != INT_MAX)
     security_limits->max_color_profile_size=max_profile_size;
   HEICSetUint64SecurityLimit(image_info,"heic:max-number-of-tiles",
